FinalCommit
Final

+++ database/migrations/2022_01_19_194113_create_scores_table.php
@@ -17,7 +17,7 @@ class CreateScoresTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('quiz_id');
             $table->unsignedBigInteger('user_id');
-            $table->double('score');
+            $table->integer('score');
             $table->timestamps();
         });
     }
